# docs/swagger.yaml
basePath: /
definitions:
...
  models.Login:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: NewPassword123!
        type: string
    type: object
...
  models.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      full_name:
        example: Jane Doe
        type: string
      password:
        example: NewPassword123!
        type: string
    type: object
...
  routes.ForgotPasswordRequest:
    properties:
      email:
        example: user@example.com
        type: string
    type: object
...
  routes.ResendOTPRequest:
    properties:
      email:
        example: user@example.com
        type: string
    type: object
  routes.ResetPasswordRequest:
    properties:
      email:
        example: user@example.com
        type: string
      newPassword:
        example: NewPassword123!
        type: string
      token:
        example: reset-token
        type: string
    type: object
...
  routes.VerifyOTPRequest:
    properties:
      email:
        example: user@example.com
        type: string
      otp:
        example: "123456"
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@ganium.ai
    name: Ganium Support
  description: Ganium AI scam detection, security investigation, scanning, authentication
    and payment API.
  license:
    name: Proprietary
  termsOfService: https://ganium.ai/terms
  title: Ganium API
  version: "1.0"
paths:
...
  /forgot-password:
    post:
      consumes:
      - application/json
      description: Starts the password recovery process for a user.
      parameters:
      - description: Email address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Forgot password
      tags:
      - Authentication
  /login:
    post:
      consumes:
      - application/json
      description: Authenticates a Ganium user and returns a JWT token.
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - Authentication
...
  /register:
    post:
      consumes:
      - application/json
      description: Creates a new Ganium user account.
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - Authentication
  /resend-otp:
    post:
      consumes:
      - application/json
      description: Resends a verification OTP to the specified email address.
      parameters:
      - description: Email address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.ResendOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Resend OTP
      tags:
      - Authentication
  /reset-password:
    post:
      consumes:
      - application/json
      description: Resets a user's password using a valid reset token.
      parameters:
      - description: Password reset details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reset password
      tags:
      - Authentication
  /verify-otp:
    post:
      consumes:
      - application/json
      description: Verifies the OTP sent to a user's email address.
      parameters:
      - description: OTP verification details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.VerifyOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify OTP
      tags:
      - Authentication
securityDefinitions:
  BearerAuth:
    description: 'JWT authentication. Enter your token using the format: Bearer <token>.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"